V10

+++ modules/PkgCreationProjet/Controllers/Base/BaseLivrableController.php
@@ -32,7 +32,7 @@ class BaseLivrableController extends AdminController
     public function index(Request $request) {
         
         $this->viewState->setContextKeyIfEmpty('livrable.index');
-
+        $this->viewState->init('filter.livrable.formateur_id'  , $this->sessionState->get('formateur_id'));
 
         // Extraire les paramètres de recherche, page, et filtres
         $livrables_params = array_merge(
@@ -47,15 +47,16 @@ class BaseLivrableController extends AdminController
         // Récupérer les statistiques et les champs filtrables
         $livrables_stats = $this->livrableService->getlivrableStats();
         $livrables_filters = $this->livrableService->getFieldsFilterable();
-
+        $livrable_instance =  $this->livrableService->createInstance();
         // Retourner la vue ou les données pour une requête AJAX
         if ($request->ajax()) {
-            return view('PkgCreationProjet::livrable._table', compact('livrables_data', 'livrables_stats', 'livrables_filters'))->render();
+            return view('PkgCreationProjet::livrable._table', compact('livrables_data', 'livrables_stats', 'livrables_filters','livrable_instance'))->render();
         }
 
-        return view('PkgCreationProjet::livrable.index', compact('livrables_data', 'livrables_stats', 'livrables_filters'));
+        return view('PkgCreationProjet::livrable.index', compact('livrables_data', 'livrables_stats', 'livrables_filters','livrable_instance'));
     }
     public function create() {
+        $this->viewState->set('scope_form.livrable.formateur_id'  , $this->sessionState->get('formateur_id'));
         $itemLivrable = $this->livrableService->createInstance();
         $natureLivrables = $this->natureLivrableService->all();
         $projets = $this->projetService->all();
@@ -90,13 +91,29 @@ class BaseLivrableController extends AdminController
         );
     }
     public function show(string $id) {
-        return $this->edit( $id);
+
+        $this->viewState->setContextKey('livrable.edit_' . $id);
+
+        $itemLivrable = $this->livrableService->find($id);
+  
+        $natureLivrables = $this->natureLivrableService->all();
+        $projets = $this->projetService->all();
+
+
+        if (request()->ajax()) {
+            return view('PkgCreationProjet::livrable._fields', compact('itemLivrable', 'natureLivrables', 'projets'));
+        }
+
+        return view('PkgCreationProjet::livrable.edit', compact('itemLivrable', 'natureLivrables', 'projets'));
+
     }
     public function edit(string $id) {
 
         $this->viewState->setContextKey('livrable.edit_' . $id);
 
         $itemLivrable = $this->livrableService->find($id);
+        $this->authorize('edit', $itemLivrable);
+
         $natureLivrables = $this->natureLivrableService->all();
         $projets = $this->projetService->all();
 
@@ -109,6 +126,9 @@ class BaseLivrableController extends AdminController
 
     }
     public function update(LivrableRequest $request, string $id) {
+        // Vérifie si l'utilisateur peut mettre à jour l'objet 
+        $livrable = $this->livrableService->find($id);
+        $this->authorize('update', $livrable);
 
         $validatedData = $request->validated();
         $livrable = $this->livrableService->update($id, $validatedData);
@@ -134,6 +154,9 @@ class BaseLivrableController extends AdminController
 
     }
     public function destroy(Request $request, string $id) {
+        // Vérifie si l'utilisateur peut mettre à jour l'objet 
+        $livrable = $this->livrableService->find($id);
+        $this->authorize('delete', $livrable);
 
         $livrable = $this->livrableService->destroy($id);
 
@@ -218,6 +241,5 @@ class BaseLivrableController extends AdminController
         ]);
     }
     
-
 
 }
